Update existing tests for Message and add extra tests
Summary: The original (and currently only) tests had gotten well out of date. This diff updates the old tests to work again. It introduces a proxy class that allows overriding the queueMessage function so that it becomes easy to convert the input Message in to an output Message and verify that the data is passed through correctly.

Test Plan: Tests pass. Passes review. CI passes

Reviewers: gpoole, einshoe, mr-asher

Reviewed By: gpoole, einshoe

Differential Revision: https://phab.adacs.org.au/D195