Add application level access to jobs
Summary:
This diff adds new functionality that allows applications (An application assigned to a specific JWT Secret Token) access to jobs that it submits itself, as well as jobs that it is configured to have access to. ie it is possible that there are 2 configured applications, application 1 can only access it's own jobs, whereas application 2 may be able to also access it's own jobs as well as jobs from application 1. This is relevant to any operations which read or operate on information about existing jobs.

Separate to application level access, is cluster level access. Each application is configured to be allowed access to specific clusters. This is only relevant to submitting new jobs.

Given that description, the code changes in this diff include:
* Application access control for
  * Job API GET (Job filtering - should only be able to return jobs an application has access to)
  * File API POST (Create file download token - should only be able to create a token for a file created by a job that an application has access to)
  * File API PATCH (Get file list for job - should only be able to retreive a file list for a job that an application has access to)

* Cluster access control for
  * Job API POST (Job submission - should only be able to submit a job to a cluster that the application has access to)

File API GET does not require any modification, if a user has a file token - they have access to the file regardless of who it is that is accessing the download.

Test Plan: Passes review. Passes CI.

Reviewers: jesmigel, gpoole, mr-asher, sosl

Reviewed By: jesmigel, mr-asher, sosl

Differential Revision: https://phab.adacs.org.au/D282