Update remote server SSH command to source a custom environment file before sourcing the virtualenv
Summary:
Previously, we were loading modules on ozstar in our bashrc file (including the python3 module). This is not acceptable, it slows SSHing down - and forces all job controller clients (and launched jobs) to use whatever version of python is loaded from bashrc. Removing the module loads breaks the remote virtualenvs, because the python interpreter the virtualenv is expecting to use is not available.

This diff simply adds an extra command to the SSH command to source a custom environment file before sourcing the virtualenv. This allows modification of the env file per instance of the job controller client. It also means that module loads don't bleed in to the launched batch jobs.

Relevant client diff: https://phab.adacs.org.au/D367

Test Plan: N/A

Reviewers: mr-asher, jesmigel

Reviewed By: jesmigel

Differential Revision: https://phab.adacs.org.au/D368