Migrate all jobs for the bilby application to bilbyui
Summary:
Since we migrated the bilby django app of gwcloud to bilbyui, we need to also migrate bilby jobs since the database search can't find any jobs with the application name bilbyui.

To deploy this, the job controller access secret config needs to be updated to also use bilbyui for the application, rather than bilby.

Test Plan: Migrates the data and bilby jobs will be discoverable again in the gwcloud UI.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D548