Fix file download id generation database multi-insert
Summary: Since the documention for correctly writing multi-row inserts in sqlpp has been updated, we can now revert the query which inserts download ids in the database to the new multi-row insert. This improves the efficiency and speed of the file POST API for large numbers of file paths per request.

Test Plan: All functionality should remain the same, and all tests should pass. A noticable increase in performance should be evident now when using large volumes of paths for the file POST API.

Reviewers: TomReichardt

Reviewed By: TomReichardt

Differential Revision: https://phab.adacs.org.au/D561