Fix file download buffering
Summary:
There existed a bug where files being transferred from client -> server -> http client would cause the entire file to buffer on the server before finally sending the file to the client last. This is not the desired behavior, as we want to keep memory usage low, and transfer latency to a minimum. Instead, the desired functionality should be that packets begin streaming to the http client as soon as packets become available, and if the transferred file buffers more than some amount (50Mb), the client should be instructed to stop streaming packets until the internal buffer has reduced in size to some amount (10Mb).

What was happening with the bug, was the HTTP response was internally buffering all packets to be sent to the client, rather than actually sending them on. This meant that the job server client just kept sending packets, since they could be queued on to the HTTP response as quickly as they could be recieved.

Now, the HTTP response correctly waits until packets have been sent before sending more, causing the internal buffering mechanism to kick in to action and work correctly.

Many gigabytes of internal file storage buffering is no longer possible, which will prevent the job controller server from crashing.

Test Plan: A new test which checks for excess memory usage during file transfers has been added.

Reviewers: TomReichardt, jesmigel

Reviewed By: jesmigel

Differential Revision: https://phab.adacs.org.au/D564