Add job deletion
Summary: So that jobs can be deleted on remote clusters, we need to implement the job deletion functionality. Here we finish up the job deletion REST api and integrate it with the cluster side. If a cluster is offline, or a job takes longer than 60 seconds to delete, then the cluster will attempt to delete the job again until the job client reports the job as deleted.

Test Plan:
New tests have been added which check the new deletion rest api, and also that test the cluster side.

Tests cover 100% of the code modified in the DELETE verb handler.
Valgrind report for tests is clean.

Reviewers: jesmigel, mr-asher, TomReichardt

Reviewed By: jesmigel, TomReichardt

Differential Revision: https://phab.adacs.org.au/D653