Fix simultanious file downloads of the same file
Summary:
There was a bug where the uuid of the file download was used for the file download map, and used all the way through to the job controller client. This lead to a weird deadlock and cluster going offline/other weirdness if the same file with the same uuid was downloaded multiple times at the same time. Instead, generating a new unique uuid that is used internally by the job controller and client per request solves this issue.

This does not change functionality, since the internal uuid can be any unique identifier, and the job download uuid is checked for validity/details before the new uuid is generated.

Test Plan: All tests pass, but uuids that are used internally are now unique for file downloads, each file download will use a unique uuid in the file download map, and in the websocket messages to fetch/transfer the file.

Reviewers: jesmigel

Reviewed By: jesmigel

Differential Revision: https://phab.adacs.org.au/D665