Update file download schema
This MR is the first part of allowing file downloads without an associate job needing to exist. Instead of having a foreign key to a job for a file download record, we have an integer field instead and include the cluster and bundle details in the file record as well.

This means the file download generation code can be refactored to allow creating a file download with an *optional* job.

MR's need to be created that update the code for creating a new file download without a job, and also for fetching a file list without a job.

Approved-by: Jesmigel Cantos <jesmigel.developer@gmail.com>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_job_server/-/merge_requests/22