Fix ClusterManager's run thread from crashing the application
Due to a previous refactor, the thread created by the start function had it's assigned variable removed because it is unused (the thread is never stopped once started). This refactor meant that the local scope thread was deconstructed when it fell out of scope, which is illegal if the thread is not joined.

Instead we now assign the thread to a static const variable that can not be deleted, but if the application was ever cleanly terminated it would be deconstructed then.

Approved-by: Jesmigel Cantos <jesmigel.developer@gmail.com>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_job_server/-/merge_requests/33