Add websocket ping/pong (To match RFC6455)
This MR implements ping/pong functionality as per RFC6455. Previously issues have been found in the proxy infrastructure in kubernetes, and firewalls between the server and connected clients when using websockets. If no traffic flowed for an arbitrary amount of time, the connection would stay established, but as soon as a packet is sent, the connection would be dropped. This MR aims to ensure that regular and reliable transmission of data is made to keep the websocket connection active and alive.

1. A new thread has been added in the cluster manager that checks for ping timeouts and issues new ping messages to clients
2. A new ping handler has been added in the websocket server that records and logs the pong timestamps.
3. Tests have been added to cover all new and updated code.

If a pong is not received in a meaningful time, we assume that the connection has dropped. We force the connection to close, and the remote client to be restarted. If a pong is received in a timely manner, then the connection is healthy, and a new ping is sent (currently) every 10 seconds.

See the following links for more details if required:-
* https://www.rfc-editor.org/rfc/rfc6455#section-5.5.2
* https://blog.martinfjordvald.com/websockets-in-nginx/
* https://websockets.readthedocs.io/en/latest/topics/timeouts.html

Approved-by: Shibli Saleheen <ssaleheen@swin.edu.au>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_job_server/-/merge_requests/35