Add Bundle DB management
Bundle DB component of the client database. This is similar to !40, but extends the database to apply to bundle operations. They are similar, but simpler, however they have an extra requirement for identification on the bundleHash. The bundleHash is sent via the client and can not be influenced by the bundle.

Approved-by: Jesmigel Cantos <jesmigel.developer@gmail.com>

Merged by: Lewis Lakerink <llakerink@swin.edu.au>

See merge request https://gitlab.com/CAS-eResearch/GWDC/code/gwcloud_job_server/-/merge_requests/41